Agrega test para prestamoService y fixea ClienteServiceTest para trabajar con persitencia en base de datos. Refactor de montos financieros a Bigdecimal.

+++ src/main/java/ar/edu/utn/frbb/tup/controller/dto/PrestamoDto.java
@@ -1,10 +1,12 @@
 package ar.edu.utn.frbb.tup.controller.dto;
 
+import java.math.BigDecimal;
+
 public class PrestamoDto {
 
     private long numeroCliente;
     private int plazoMeses;
-    private double montoPrestamo;
+    private BigDecimal montoPrestamo;
     private String moneda;
 
 
@@ -24,11 +26,11 @@ public class PrestamoDto {
         this.plazoMeses = plazoMeses;
     }
 
-    public double getMontoPrestamo() {
+    public BigDecimal getMontoPrestamo() {
         return montoPrestamo;
     }
 
-    public void setMontoPrestamo(double montoPrestamo) {
+    public void setMontoPrestamo(BigDecimal montoPrestamo) {
         this.montoPrestamo = montoPrestamo;
     }
